src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
albumentations=0.5.2=py_0
blas=1.0=mkl
brotlipy=0.7.0=py38h27cfd23_1003
ca-certificates=2021.4.13=h06a4308_1
certifi=2020.12.5=py38h06a4308_0
cffi=1.14.3=py38h261ae71_2
chardet=3.0.4=py38h06a4308_1003
cloudpickle=1.6.0=py_0
conda=4.10.1=py38h06a4308_1
conda-package-handling=1.7.2=py38h03888b9_0
cryptography=3.2.1=py38h3c74f83_1
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=py38_0
cytoolz=0.11.0=py38h7b6447c_0
dask-core=2021.4.1=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=pyhd3eb1b0_0
efficientnet-pytorch=0.6.3=pyh9f0ad1d_0
expat=2.3.0=h2531618_2
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
fsspec=0.9.0=pyhd3eb1b0_0
geos=3.8.0=he6710b0_0
glib=2.68.1=h36276a3_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=2.10=py_0
imageio=2.9.0=pyhd3eb1b0_0
imgaug=0.4.0=py_0
intel-openmp=2021.2.0=h06a4308_610
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=py38h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.5.0=h14aa051_19
libgfortran4=7.5.0=h14aa051_19
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
locket=0.2.1=py38h06a4308_1
lz4-c=1.9.3=h2531618_0
matplotlib=3.3.4=py38h06a4308_0
matplotlib-base=3.3.4=py38h62a2d02_0
mkl=2021.2.0=h06a4308_296
mkl-service=2.3.0=py38h27cfd23_1
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.1=py38ha9443f7_2
munch=2.5.0=py_0
ncurses=6.2=he6710b0_1
networkx=2.5.1=pyhd3eb1b0_0
ninja=1.10.2=hff7bd54_1
numpy=1.20.1=py38h93e21f0_0
numpy-base=1.20.1=py38h7d8b39e_0
olefile=0.46=py_0
opencv-python-headless=4.5.2.52=py38_0
openssl=1.1.1k=h27cfd23_0
partd=1.2.0=pyhd3eb1b0_0
pcre=8.44=he6710b0_0
pillow=8.2.0=py38he98fc37_0
pip=21.0.1=py38h06a4308_0
pretrainedmodels=0.7.4=pyhd8ed1ab_2
pycosat=0.6.3=py38h7b6447c_1
pycparser=2.20=py_2
pyopenssl=19.1.0=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pysocks=1.7.1=py38h06a4308_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.8=1_cp38
pytorch=1.6.0=py3.8_cuda10.1.243_cudnn7.6.3_0
pywavelets=1.1.1=py38h7b6447c_2
pyyaml=5.4.1=py38h27cfd23_1
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
requests=2.24.0=py_0
ruamel_yaml=0.15.87=py38h7b6447c_1
scikit-image=0.18.1=py38ha9443f7_0
scipy=1.6.2=py38had2a1c9_1
segmentation-models-pytorch=0.1.3=pyhd8ed1ab_1
setuptools=50.3.1=py38h06a4308_1
shapely=1.7.1=py38h98ec03d_0
sip=4.19.13=py38he6710b0_0
six=1.15.0=py38h06a4308_0
sqlite=3.33.0=h62c20be_0
tifffile=2020.10.1=py38hdd07704_2
timm=0.3.2=pyhd8ed1ab_0
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torchsummary=1.5.1=py38_0
torchvision=0.7.0=py38_cu101
tornado=6.1=py38h27cfd23_0
tqdm=4.51.0=pyhd3eb1b0_0
urllib3=1.25.11=py_0
wheel=0.35.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
